+++ shared/src/main/scala/com/raquo/waypoint/package.scala
@@ -1,7 +1,7 @@
 package com.raquo
 
 import urldsl.errors._
-import urldsl.vocabulary.UrlMatching
+import urldsl.vocabulary.{PathQueryFragmentMatching, UrlMatching}
 
 /** This is the public API. Import com.raquo.waypoint or com.raquo.waypoint._ */
 package object waypoint extends Waypoint(
@@ -10,8 +10,15 @@ package object waypoint extends Waypoint(
   DummyError.dummyErrorIsFragmentMatchingError
 ) {
 
+  // @TODO[URL-DSL] PatternArgs should be FragmentPatternArgs with FragmentArgs=Unit, not different types
+
   /** A bundle of path args and query param args */
   type PatternArgs[PathArgs, QueryArgs] = UrlMatching[PathArgs, QueryArgs]
 
+  /** A bundle of path args, query param args, and fragment args */
+  type FragmentPatternArgs[PathArgs, QueryArgs, FragmentArgs] = PathQueryFragmentMatching[PathArgs, QueryArgs, FragmentArgs]
+
   @inline def PatternArgs: UrlMatching.type = UrlMatching
+
+  @inline def FragmentPatternArgs: PathQueryFragmentMatching.type = PathQueryFragmentMatching
 }